# dispatch speedups