--- a/base/serialize.jl
+++ b/base/serialize.jl
@@ -956,6 +956,9 @@ function deserialize_typename(s::AbstractSerializer, number)
         maxa = deserialize(s)::Int
         if makenew
             tn.mt = ccall(:jl_new_method_table, Any, (Any, Any), name, tn.module)
+            if isempty(parameters)
+                tn.mt.offs = 1
+            end
             tn.mt.name = mtname
             tn.mt.max_args = maxa
             for def in defs