--- a/src/datatype.c
+++ b/src/datatype.c
@@ -46,6 +46,7 @@ JL_DLLEXPORT jl_methtable_t *jl_new_method_table(jl_sym_t *name, jl_module_t *mo
     mt->kwsorter = NULL;
     mt->backedges = NULL;
     JL_MUTEX_INIT(&mt->writelock);
+    mt->offs = 0;
     return mt;
 }
 
@@ -395,6 +396,8 @@ JL_DLLEXPORT jl_datatype_t *jl_new_datatype(
             if (!abstract) {
                 tn->mt = jl_new_method_table(name, module);
                 jl_gc_wb(tn, tn->mt);
+                if (jl_svec_len(parameters) == 0)
+                    tn->mt->offs = 1;
             }
         }
         t->name = tn;