--- a/src/gf.c
+++ b/src/gf.c
@@ -123,13 +123,13 @@ void jl_call_tracer(tracer_cb callback, jl_value_t *tracee)
 /// ----- Definitions for various internal TypeMaps ----- ///
 
 const struct jl_typemap_info method_defs = {
-    0, &jl_method_type
+    0, 0, &jl_method_type
 };
 const struct jl_typemap_info lambda_cache = {
-    0, &jl_method_instance_type
+    0, 1, &jl_method_instance_type
 };
 const struct jl_typemap_info tfunc_cache = {
-    1, &jl_any_type
+    1, 0, &jl_any_type
 };
 
 static int8_t jl_cachearg_offset(jl_methtable_t *mt)