@KristofferC 2017-05-09
Would these benchmarks be suitable to add to Nanosoldier?