@JeffBezanson 2017-05-09
Yes, I don't see why not. Let's add them.