@nanosoldier 2017-05-09
[Your benchmark job](https://github.com/JuliaLang/julia/pull/21760) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/2cc4084d858b6c955012d089a0b9ad19737f67b8/a849c1c_vs_d214d57/report.md). cc @jrevels