@vtjnash 2017-05-17
This feels unlikely to be valid, but I'll have to work a bit harder to understand it to see if I can create a good counterexample.