@JeffBezanson 2017-05-17
It actually ended up being really simple. First I restored using either 0 or 1 as the offset based on whether the function type has parameters. Hopefully that's uncontroversial.

Next, I observed that the `->any` table is never used for the 0th argument, because you can't have `Any` in that slot. So I use it to split `Type{T}`s based on whether T is_cache_leaf. If it is, it gets inserted as before starting with offset 0. Otherwise, the offset is incremented to 1 and it's inserted into the `->any` table. Both are checked on lookup (as before). This concept seems sound to me.