@vtjnash 2017-05-17
> because you can't have Any in that slot

This is true, but actually irrelevant. The question is whether qualifying for the split guarantees it is more specific than anything in the sorted list. The `offset = 0` test also seems to be a red-herring – it likely should be valid regardless of `offset`. If that's true, you can just introduce this universally as a new field. I think we also need to double-check that this will handle Varargs correctly.