@KristofferC 2017-05-26
Bump, would be nice to get this in 0.6 if it is determined to be sound.