@JeffBezanson 2017-07-01
@vtjnash bump. Maybe we can put this on master and see how it goes?