@vtjnash 2017-07-01
I'm still fairly certain that aaaf9f4ff54dbec24eacd7d194c8c474a1277429 violates specificity ordering. The other two commits look ok.

> I'm not sure we want to add another field

Since you're using one field to mean two different things, the question isn't whether we add another field to disambiguate those, it's whether it would be _valid_ to remove the `offs == 0` test (even if we end up deciding to keep that condition in place). If the system still works without that condition, then that commit may be OK. Otherwise that commit is likely invalid.