@JeffBezanson 2017-07-01
I'll separate the two safer commits and then explore this.