@JeffBezanson 2017-07-11
Ok, I came up with a different approach that's simpler and more general. I changed the meaning of the `any` field: instead of using it to skip slots equal to `Any`, it's used to skip all non-leaf slots *if* some later slot is a cache_leaf. This means each level is neatly split into three cases:

1. Leaf at this offset; use hash splitting.
2. Leaf at some later offset; use `any`.
3. No remaining leaf types, use `linear`.

Those cases should be correctly ordered, most specific first. If this checks out, the field should perhaps be renamed `skip`.