@JeffBezanson 2017-07-12
Just look at the newest commit (`speed up 0-arg constructor dispatch`). The others are unchanged, and that one is all new.

I think the solution is to use this trick only for unsorted maps. Method caches are far, far more performance-critical than sorted method lists, and I highly doubt they will ever be able to handle anything as complex as `(a::ConjArray{T,N,A}, i::Vararg{Int64,N}) where {T, N}`.