@JeffBezanson 2017-07-14
@vtjnash : I removed the change to the `unsorted` flag in `lambda_cache`, and added the flag I described. What do you think? This speedup would be nice...