@vtjnash 2017-10-12
Bump? I agree, this speedup _would_ be nice :)