@StefanKarpinski 2017-10-12
Unless this is blocking something, I think we should wait until post 1.0 to merge this.